makes relevant changes
